Extract SDK executor plugin runtime wiring
